requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
av==10.0.0
basicsr==1.4.2
boto3==1.35.81
botocore==1.35.81
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
ctranslate2==3.24.0
cycler==0.12.1
decorator==4.4.2
exceptiongroup==1.2.2
executing==2.1.0
facexlib==0.3.0
faster-whisper==0.7.0
ffmpeg-python==0.2.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
future==1.0.0
gfpgan==1.3.8
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-texttospeech==2.22.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.18.1
ipython-autotime==0.3.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
moviepy==2.0.0.dev2
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
onnxruntime==1.19.2
openai-whisper==20231117
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
srt==3.5.3
stack-data==0.6.3
sympy==1.13.1
tb-nightly==2.19.0a20241006
tensorboard-data-server==0.7.2
tifffile==2024.8.30
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.0.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.47.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.0.4
yapf==0.40.2
zipp==3.20.2
